update benchmark files, use latest version of phpbench
